SWM-4537 - Separated BuildConfig into BuildConfig and RuntimeConfig
Watchtower has two modes of operation:
build (so it can be run later, probably not by watchtower)
run (starting/hosting the server, normally for dev, or stats)

Different information is available at each times, for example when running the app from the /dist folder, watchtower can no longer access the configuration while created the build. Meaning it is actually guessing where the assets live.
During watchtower build it will create a watchtower.config file in the output folder, this then is used to create the runtime config when the server is started not in dev mode