Adds prebody page tag to insert content before root element
We have run into the need to be able to insert content before the root
element or another way of thinking about it, as the first element in the
body.

This new prop is called preBody and is by default empty but you can add
html content as needed. Tests have been added in ssr.test.tsx to test all
of the page tags.